some modify

+++ CCEH/src/CCEH_buflog.h
@@ -1,17 +1,18 @@
 #ifndef CCEH_H_
 #define CCEH_H_
 
+#include <libpmemobj.h>
+#include <pthread.h>
+
+#include <cmath>
+#include <cstdlib>
 #include <cstring>
 #include <vector>
-#include <cmath>
-#include <vector>
-#include <cstdlib>
-#include <pthread.h>
-#include <libpmemobj.h>
-#include "util.h"
-#include "../../src/buflog.h"
 
-#define TOID_ARRAY(x) TOID(x)
+#include "../../src/buflog.h"
+#include "util.h"
+
+#define TOID_ARRAY(x) TOID (x)
 
 typedef size_t Key_t;
 typedef const char* Value_t;
@@ -20,7 +21,7 @@ const Key_t SENTINEL = -2;
 const Key_t INVALID = -1;
 const Value_t NONE = 0x0;
 
-struct Pair{
+struct Pair {
     Key_t key;
     Value_t value;
 };
@@ -28,15 +29,15 @@ struct Pair{
 class CCEH;
 struct Directory;
 struct Segment;
-POBJ_LAYOUT_BEGIN(HashTable);
-POBJ_LAYOUT_ROOT(HashTable, CCEH);
-POBJ_LAYOUT_TOID(HashTable, struct Directory);
-POBJ_LAYOUT_ROOT(HashTable, struct Segment);
-POBJ_LAYOUT_TOID(HashTable, TOID(struct Segment));
-POBJ_LAYOUT_END(HashTable);
+POBJ_LAYOUT_BEGIN (HashTable);
+POBJ_LAYOUT_ROOT (HashTable, CCEH);
+POBJ_LAYOUT_TOID (HashTable, struct Directory);
+POBJ_LAYOUT_ROOT (HashTable, struct Segment);
+POBJ_LAYOUT_TOID (HashTable, TOID (struct Segment));
+POBJ_LAYOUT_END (HashTable);
 
 constexpr size_t kSegmentBits = 8;
-constexpr size_t kMask = (1 << kSegmentBits)-1;
+constexpr size_t kMask = (1 << kSegmentBits) - 1;
 constexpr size_t kShift = kSegmentBits;
 constexpr size_t kSegmentSize = (1 << kSegmentBits) * 16 * 4;
 constexpr size_t kWriteBufferSize = kSegmentSize / 2 / 256;
@@ -45,163 +46,160 @@ constexpr size_t kNumCacheLine = 8;
 constexpr size_t kCuckooThreshold = 16;
 
 using WriteBuffer = buflog::WriteBuffer<kWriteBufferSize>;
-//constexpr size_t kCuckooThreshold = 32;
+// constexpr size_t kCuckooThreshold = 32;
 
-struct Segment{
-    static const size_t kNumSlot = kSegmentSize/sizeof(Pair);
+struct Segment {
+    static const size_t kNumSlot = kSegmentSize / sizeof (Pair);
 
-    Segment(void){ }
-    ~Segment(void){ }
+    Segment (void) {}
+    ~Segment (void) {}
 
-    void initSegment(void){
-		for(int i=0; i<kNumSlot; ++i){
-			bucket[i].key = INVALID;
-		}
-		local_depth = 0;
-		sema = 0;
-		bufnode_ = new WriteBuffer();
+    void initSegment (void) {
+        for (size_t i = 0; i < kNumSlot; ++i) {
+            bucket[i].key = INVALID;
+        }
+        local_depth = 0;
+        sema = 0;
+        bufnode_ = new WriteBuffer ();
     }
 
-    void initSegment(size_t depth){
-		for(int i=0; i<kNumSlot; ++i){
-			bucket[i].key = INVALID;
-		}
-		local_depth = depth;
-		sema = 0;
-		bufnode_ = new WriteBuffer(depth);
+    void initSegment (size_t depth) {
+        for (size_t i = 0; i < kNumSlot; ++i) {
+            bucket[i].key = INVALID;
+        }
+        local_depth = depth;
+        sema = 0;
+        bufnode_ = new WriteBuffer (depth);
     }
 
-    bool suspend(void){
-	int64_t val;
-	do{
-	    val = sema;
-	    if(val < 0)
-		return false;
-	}while(!CAS(&sema, &val, -1));
+    bool suspend (void) {
+        int64_t val;
+        do {
+            val = sema;
+            if (val < 0) return false;
+        } while (!CAS (&sema, &val, -1));
 
-	int64_t wait = 0 - val - 1;
-	while(val && sema != wait){
-	    asm("nop");
-	}
-	return true;
+        int64_t wait = 0 - val - 1;
+        while (val && sema != wait) {
+            asm("nop");
+        }
+        return true;
     }
 
-    bool lock(void){
-	int64_t val = sema;
-	while(val > -1){
-	    if(CAS(&sema, &val, val+1))
-		return true;
-	    val = sema;
-	}
-	return false;
+    bool lock (void) {
+        int64_t val = sema;
+        while (val > -1) {
+            if (CAS (&sema, &val, val + 1)) return true;
+            val = sema;
+        }
+        return false;
     }
 
-    void unlock(void){
-	int64_t val = sema;
-	while(!CAS(&sema, &val, val-1)){
-	    val = sema;
-	}
+    void unlock (void) {
+        int64_t val = sema;
+        while (!CAS (&sema, &val, val - 1)) {
+            val = sema;
+        }
     }
 
-    int Insert(PMEMobjpool*, Key_t&, Value_t, size_t, size_t);
-    bool Insert4split(Key_t&, Value_t, size_t);
-    TOID(struct Segment)* Split(PMEMobjpool*);
-	struct Segment* SplitDram(WriteBuffer::Iterator& iter);
-    std::vector<std::pair<size_t, size_t>> find_path(size_t, size_t);
-    void execute_path(PMEMobjpool*, std::vector<std::pair<size_t, size_t>>&, Key_t&, Value_t);
-    void execute_path(std::vector<std::pair<size_t, size_t>>&, Pair);
-    size_t numElement(void);
+    int Insert (PMEMobjpool*, Key_t&, Value_t, size_t, size_t);
+    bool Insert4split (Key_t&, Value_t, size_t);
+    TOID (struct Segment) * Split (PMEMobjpool*);
+    struct Segment* SplitDram (WriteBuffer::Iterator& iter);
+    std::vector<std::pair<size_t, size_t>> find_path (size_t, size_t);
+    void execute_path (PMEMobjpool*, std::vector<std::pair<size_t, size_t>>&, Key_t&, Value_t);
+    void execute_path (std::vector<std::pair<size_t, size_t>>&, Pair);
+    size_t numElement (void);
 
-    Pair bucket[kNumSlot];    
-	int64_t sema = 0;
-    size_t local_depth;	
-	WriteBuffer* bufnode_;
+    Pair bucket[kNumSlot];
+    int64_t sema = 0;
+    size_t local_depth;
+    WriteBuffer* bufnode_;
 };
 
-struct Directory{
+struct Directory {
     static const size_t kDefaultDepth = 10;
 
-    TOID_ARRAY(TOID(struct Segment)) segment;	
+    TOID_ARRAY (TOID (struct Segment)) segment;
     int64_t sema = 0;
-    size_t capacity;		
-    size_t depth;	
+    size_t capacity;
+    size_t depth;
 
-    bool suspend(void){
-	int64_t val;
-	do{
-	    val = sema;
-	    if(val < 0)
-		return false;
-	}while(!CAS(&sema, &val, -1));
+    bool suspend (void) {
+        int64_t val;
+        do {
+            val = sema;
+            if (val < 0) return false;
+        } while (!CAS (&sema, &val, -1));
 
-	int64_t wait = 0 - val - 1;
-	while(val && sema != wait){
-	    asm("nop");
-	}
-	return true;
+        int64_t wait = 0 - val - 1;
+        while (val && sema != wait) {
+            asm("nop");
+        }
+        return true;
     }
 
-    bool lock(void){
-	int64_t val = sema;
-	while(val > -1){
-	    if(CAS(&sema, &val, val+1))
-		return true;
-	    val = sema;
-	}
-	return false;
+    bool lock (void) {
+        int64_t val = sema;
+        while (val > -1) {
+            if (CAS (&sema, &val, val + 1)) return true;
+            val = sema;
+        }
+        return false;
     }
 
-    void unlock(void){
-	int64_t val = sema;
-	while(!CAS(&sema, &val, val-1)){
-	    val = sema;
-	}
+    void unlock (void) {
+        int64_t val = sema;
+        while (!CAS (&sema, &val, val - 1)) {
+            val = sema;
+        }
     }
 
-    Directory(void){ }
-    ~Directory(void){ }
+    Directory (void) {}
+    ~Directory (void) {}
 
-    void initDirectory(void){
-	depth = kDefaultDepth;
-	capacity = pow(2, depth);
-	sema = 0;
-	INFO("Directory capacity: %lu. depth %lu\n", capacity, depth);
-	printf("Directory capacity: %lu. depth %lu\n", capacity, depth);
+    void initDirectory (void) {
+        depth = kDefaultDepth;
+        capacity = pow (2, depth);
+        sema = 0;
+        INFO ("Directory capacity: %lu. depth %lu\n", capacity, depth);
+        printf ("Directory capacity: %lu. depth %lu\n", capacity, depth);
     }
 
-    void initDirectory(size_t _depth){
-		depth = _depth;
-		capacity = pow(2, _depth);
-		INFO("Directory capacity: %lu. depth %lu\n", capacity, depth);
-		printf("Directory capacity: %lu. depth %lu\n", capacity, depth);
-		sema = 0;
+    void initDirectory (size_t _depth) {
+        depth = _depth;
+        capacity = pow (2, _depth);
+        INFO ("Directory capacity: %lu. depth %lu\n", capacity, depth);
+        printf ("Directory capacity: %lu. depth %lu\n", capacity, depth);
+        sema = 0;
     }
 };
 
-class CCEH{
-    public:
-	CCEH(void){ }
-	~CCEH(void){ }
-	void initCCEH(PMEMobjpool*);
-	void initCCEH(PMEMobjpool*, size_t);
+class CCEH {
+public:
+    CCEH (void) {}
+    ~CCEH (void) {}
+    void initCCEH (PMEMobjpool*);
+    void initCCEH (PMEMobjpool*, size_t);
 
-	void Insert(PMEMobjpool*, Key_t&, Value_t);
-	void insert(PMEMobjpool*, Key_t&, Value_t, bool with_lock);
-	void mergeBufAndSplitWhenNeeded(PMEMobjpool*, WriteBuffer* bufnode, Segment_toid& target, size_t x);
-	bool InsertOnly(PMEMobjpool*, Key_t&, Value_t);
-	bool Delete(Key_t&);
-	Value_t Get(Key_t&);
-	Value_t get(Key_t&);
-	Value_t FindAnyway(Key_t&);
+    void Insert (PMEMobjpool*, Key_t&, Value_t);
+    void insert (PMEMobjpool*, Key_t&, Value_t, bool with_lock);
+    void mergeBufAndSplitWhenNeeded (PMEMobjpool*, WriteBuffer* bufnode, Segment_toid& target,
+                                     size_t x);
+    bool InsertOnly (PMEMobjpool*, Key_t&, Value_t);
+    bool Delete (Key_t&);
+    Value_t Get (Key_t&);
+    Value_t get (Key_t&);
+    Value_t FindAnyway (Key_t&);
 
-	double Utilization(void);
-	size_t Capacity(void);
-	void Recovery(PMEMobjpool*);
+    double Utilization (void);
+    size_t Capacity (void);
+    void Recovery (PMEMobjpool*);
 
-	bool crashed = true;
-    private:
-	TOID(struct Directory) dir;
+    bool crashed = true;
+
+private:
+    TOID (struct Directory) dir;
 };
 
 #endif
-
